+++ src/main/java/org/truffleruby/core/module/ModuleNodes.java
@@ -1049,14 +1049,18 @@ public abstract class ModuleNodes {
         @TruffleBoundary
         @Specialization(guards = "isRubyUnboundMethod(method)")
         public DynamicObject defineMethod(DynamicObject module, String name, DynamicObject method, NotProvided block) {
-            final DynamicObject origin = Layouts.UNBOUND_METHOD.getOrigin(method);
-            if (!ModuleOperations.canBindMethodTo(origin, module)) {
-                throw new RaiseException(getContext(), coreExceptions().typeError("bind argument must be a subclass of " + Layouts.MODULE.getFields(origin).getName(), this));
+            final InternalMethod internalMethod = Layouts.UNBOUND_METHOD.getMethod(method);
+            if (!ModuleOperations.canBindMethodTo(internalMethod, module)) {
+                final DynamicObject declaringModule = internalMethod.getDeclaringModule();
+                if (RubyGuards.isSingletonClass(declaringModule)) {
+                    throw new RaiseException(getContext(), coreExceptions().typeError(
+                            "can't bind singleton method to a different class", this));
+                } else {
+                    throw new RaiseException(getContext(), coreExceptions().typeError("bind argument must be a subclass of " + Layouts.MODULE.getFields(declaringModule).getName(), this));
+                }
             }
 
-            // TODO CS 5-Apr-15 TypeError if the method came from a singleton
-
-            return addMethod(module, name, Layouts.UNBOUND_METHOD.getMethod(method));
+            return addMethod(module, name, internalMethod);
         }
 
         @TruffleBoundary