+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Bug fixes:
 * Fixed `Symbol#match` returning `MatchData` (#1706).
 * Allow `Time#strftime` to be called with binary format strings.
 * Do not modify the argument passed to `IO#write` when the encoding does not match (#1714).
+* Use the class where the method was defined to check if an `UnboundMethod` can be used for `#define_method` (#1710).
 
 Compatibility: